+++ spec/clustering/VectorWithNorm_spec.lua
@@ -10,7 +10,7 @@ describe('clustering.VectorWithNorm', function()
     local values = {0.1, 0.3, 4}
     local norm = 2
     local vecWithNorm = VectorWithNorm:new(values, norm)
-    local expected = '([0.1,0.3,4],2)'
+    local expected = '({0.1,0.3,4},2)'
     local actual = tostring(vecWithNorm)
     assert.equal(expected, actual)
   end)
@@ -20,7 +20,7 @@ describe('clustering.VectorWithNorm', function()
     local norm = 2
     local vec = Vectors.dense(values)
     local vecWithNorm = VectorWithNorm:new(vec, norm)
-    local expected = '([0.1,0.3,4],2)'
+    local expected = '({0.1,0.3,4},2)'
     local actual = tostring(vecWithNorm)
     assert.equal(expected, actual)
   end)
@@ -31,18 +31,18 @@ describe('clustering.VectorWithNorm', function()
     local norm = 2
     local vec = Vectors.sparse(3, indices, values)
     local vecWithNorm = VectorWithNorm:new(vec, norm)
-    local expected = '((3,[1,3,4],[0.1,0.3,4]),2)'
+    local expected = '((3,{1,3,4},{0.1,0.3,4}),2)'
     local actual = tostring(vecWithNorm)
     assert.equal(expected, actual)
   end)
   
   it('converts to DenseVector', function()
-    local indices = {1, 3, 4}
+    local indices = {0, 2, 3}
     local values = {0.1, 0.3, 4}
     local norm = 3
     local vec = Vectors.sparse(3, indices, values)
     local vecWithNorm = VectorWithNorm:new(vec, norm):toDense()
-    local expected = '([0.1,0,0.3,4],3)'
+    local expected = '({0.1,0,0.3,4},3)'
     local actual = tostring(vecWithNorm)
     assert.equal(expected, actual)
   end)