+++ src/stuart-ml/linalg/DenseVector.lua
@@ -28,7 +28,7 @@ function DenseVector:__index(key)
 end
 
 function DenseVector:__tostring()
-  return '[' .. table.concat(self.values,',') .. ']'
+  return '{' .. table.concat(self.values,',') .. '}'
 end
 
 function DenseVector:argmax()
@@ -53,7 +53,7 @@ end
 
 function DenseVector:foreachActive(f)
   for i,value in ipairs(self.values) do
-    f(i, value)
+    f(i-1, value)
   end
 end