+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Changelog
+
+## [Unreleased]
+### Added
+- `linalg:` Ported `BLAS` `axpy()` and `dot()`
+- `linalg:` Ported `Vectors` `sqdist()`
+- `util:` Ported `MLUtils` `EPSILON` and `fastSquaredDistance()`
+
+### Changed
+- `linalg: ` Vector types are now 0-based like Apache Spark
+
 ## [0.1.0] - 2017-10-28
 ### Added
 - `clustering:` Ported `VectorWithNorm` datatype